+++ src/main/java/de/autohaus/logic/DeleteMotor.java
@@ -8,14 +8,21 @@ import static de.autohaus.model.Connect.connect;
 
 public class DeleteMotor {
 
-    public DeleteMotor(JTextField id) {
-        try {
-            String sql = "DELETE FROM motor WHERE MTID = ?";
-            PreparedStatement psm = connect().prepareStatement(sql);
-            psm.setString(1, id.getText());
-            psm.executeUpdate();
-        } catch (SQLException ex) {
-            ex.printStackTrace();
+    public DeleteMotor(JTextField id, String benutzerID, JTextField MTtextFieldMTID) {
+        if (!id.getText().equals("0") && !id.getText().isBlank() && !id.getText().isEmpty()) {
+            try {
+                String sql = "DELETE FROM motor WHERE MTID = ?";
+                PreparedStatement psm = connect().prepareStatement(sql);
+                psm.setString(1, id.getText());
+                psm.executeUpdate();
+
+                new InsertLogs(benutzerID, "geloescht", "Motor", MTtextFieldMTID.getText());
+            } catch (SQLException ex) {
+                new InsertLogs(benutzerID, "geloescht", "Motor", MTtextFieldMTID.getText(), true);
+                ex.printStackTrace();
+            }
+        } else {
+            new InsertLogs(benutzerID, "geloescht", "Motor", true);
         }
     }
